시작 번호에 맞게 작업 툴 프레임 계산되도록 변경

+++ Inspection/Inspection_post.py
@@ -4,26 +4,89 @@ import Label
 import shutil
 import re
 
-# G:\내 드라이브\PM2팀\#김한솔\23-DS-041_현대자동차\3. 데이터\2. 작업 데이터\02_Inspection\01_IRIS_JX013\03_Urban\HKMC-N2202209-240207
+# G:/내 드라이브/PM2팀/#김한솔/23-DS-041_현대자동차/3. 데이터/2. 작업 데이터/02_Inspection/01_IRIS_JX013/03_Urban\HKMC-N2202209-240207
+
+def process_sequence(data_path, sequence, save_path):
+    sequence_path = os.path.join(data_path, sequence)
+    print(sequence_path)
+    parsing_num_txt = 'parsing_num.txt'
+    f = open(os.path.join(sequence_path,parsing_num_txt), 'r')
+    lines = f.read()
+    start_frame = int(lines.split('~')[0])
+
+    csv_save_path = process_excel_files(sequence_path, save_path,start_frame)
+    cel = csv_error_list(path= csv_save_path)
+
+    iel = copy_and_rename_images(sequence_path, save_path, sequence)
+    sorted_iel = sorted(iel, key=lambda x: (int(x.split('_')[0]), int(x.split('_')[1])))
+
+    error_show = compare_lists(sorted_iel, cel)
+    print(error_show)
+    if error_show:
+        raise Exception(f"에러 개수 불일치 {error_show}")
+    else:
+        print("에러 개수 일치")
+
+
+
+
+def compare_lists(iel, cel):
+    if len(iel) != len(cel):
+        if len(iel) > len(cel):
+            long, short = iel, cel
+            print(f'image 폴더에 오류가 더 많음')
+        else:
+            long, short = cel, iel
+            print(f'csv 파일에 오류가 더 많음')
+        for i in range(len(short)):
+            if short[i] in long:
+                long.remove(short[i])
+        return long
+
+
+
+def image_error_list(iel):
+    result = []
+    for i in iel:
+        if i.endswith('.ini'):
+            continue
+        val = i.split('LDR-1_')[1][:-4]
+        result.append(val)
+    return result
+
+def transform(row):
+    frame_number = int(row['Frame Index'].replace('.pcd', '').lstrip('0'))  # 앞의 0 제거하고 숫자만 남김
+    inspection_index = row['Inspection Index']
+    return f"{frame_number}_{inspection_index}"
+def csv_error_list(path):
+    result = []
+    df = pd.read_csv(path)
+    df['el'] = df.apply(transform, axis=1)
+    result = df['el'].tolist()
+    return result
 
 def copy_and_rename_images(sequence_path, save_path, sequence):
     for value in os.listdir(sequence_path):
         value_path = os.path.join(sequence_path, value)
-
         if os.path.isdir(value_path) and value.startswith('LDR'):
             dest_path = os.path.join(save_path, value)
-            shutil.copytree(value_path, dest_path)
+            if not os.path.exists(dest_path):
+                shutil.copytree(value_path, dest_path)
+                for image in os.listdir(dest_path):
+                    if image.endswith('.jpg') or image.endswith('.JPG'):
+                        if image.endswith('.JPG'):
+                            image.replace('.JPG', '.jpg')
+                        new_image_name = image.replace('-', '_')
+                        parts = new_image_name.split('_')
+                        number_part = parts[-2].zfill(6)
+                        index_part = parts[-1]
+                        new_image_name = f'LDR_INSPECTION-{sequence}_' + '_'.join(
+                            parts[:-2]) + f'{number_part}_{index_part}'
+                        os.rename(os.path.join(dest_path, image), os.path.join(dest_path, new_image_name))
 
-
-            # Rename jpg files in the copied directory
-            for image in os.listdir(dest_path):
-                if image.endswith('.jpg'):
-                    new_image_name = image.replace('-', '_')
-                    new_image_name = f'LDR_INSPECTION-{sequence}_{new_image_name}'
-                    os.rename(os.path.join(dest_path, image), os.path.join(dest_path, new_image_name))
-
-            return len(os.listdir(dest_path)) - 1
-def process_excel_files(sequence_path, save_path):
+            iel = image_error_list(os.listdir(dest_path))
+            return iel
+def process_excel_files(sequence_path, save_path, start_frame):
     for value in os.listdir(sequence_path):
         if value.endswith('.xlsx'):
             value_path = os.path.join(sequence_path, value)
@@ -35,8 +98,9 @@ def process_excel_files(sequence_path, save_path):
             df['Selected Point Number'] = df['Selected Points Class/Object'].str.extract(r'\[(-?\d+)\]')
 
             # Group by Frame Number
-            start_value = df['Frame Number'].min() - 50
-            df['annotations'] = ((df['Frame Number'] - start_value) // 50).astype(int) - 1
+            start_value = start_frame
+            print(start_value)
+            df['annotations'] = ((df['Frame Number'] - start_value) // 50).astype(int)
 
             df_to_dict = df.groupby('annotations')['Selected Point Number'].apply(list).to_dict()
 
@@ -62,7 +126,8 @@ def process_excel_files(sequence_path, save_path):
 
                     for id_str in df_to_dict[filenum]:
                         id_int = int(id_str)
-                        if id_int > 0:
+                        print(annotation, id_int)
+                        if id_int >= 0:
                             necessary_vertex.append(box_vertices_list[id_int].tolist())
                         else:
                             necessary_vertex.append(miss_gt_box_vertices[-id_int - 1].tolist())
@@ -79,63 +144,44 @@ def process_excel_files(sequence_path, save_path):
             df.to_csv(os.path.join(save_path, name), index=False)
             print(f'{name} Done')
 
-            return df.shape[0]
-def process_sequence(data_path, sequence, save_path):
-    sequence_path = os.path.join(data_path, sequence)
-
-    cnt_image_error = copy_and_rename_images(sequence_path, save_path, sequence)
-    cnt_csv_error = process_excel_files(sequence_path, save_path)
-    print(f"이미지 : {cnt_image_error}\n"
-          f"csv : {cnt_csv_error}")
-
-    raise Exception("에러 개수 불일치") if cnt_image_error != cnt_csv_error else print("에러 개수 일치")
-
-
-
-
-
-# def check_result(inspection_dir):
-#     for value in os.listdir(inspection_dir):
-#         if value.endswith('.csv'):
-#             df = pd.read_csv(os.path.join(inspection_dir, value))
-#             df['Combined'] = df['Frame Index'].str.replace('.pcd', '').astype(str) + '_' + df[
-#                 'Inspection Index'].astype(str)
-#             error_list = df['Combined'].tolist()
-#             normalized_list1 = [re.sub(r'^0+', '', item) for item in error_list]
-#             for dir in os.listdir(inspection_dir):
-#                 if dir == value.split('.csv')[0]:
-#                     print(value.split('.csv')[0])
-#                     images_dir = os.path.join(inspection_dir, dir)
-#                     images_dir_list = os.listdir(images_dir)
-#                     normalized_list2 = [re.sub(r'\.jpg$', '', re.sub(r'^0+', '', item)) for item in images_dir_list if
-#                                         not item.endswith('.ini')]
-#
-#             sorted_list1 = sorted(normalized_list1, key=lambda x: (int(x.split('_')[0]), int(x.split('_')[1])))
-#             sorted_list2 = sorted(normalized_list2, key=lambda x: (int(x.split('_')[0]), int(x.split('_')[1])))
-#             print(len(sorted_list1), sorted_list1)
-#             print(len(sorted_list2), sorted_list2)
-
+            return os.path.join(save_path, name)
 
 if __name__ == '__main__':
     bucket_name = "coop-selectstar-7000527-241231/"
+    ''' :parameter
+        step
+            :param
+            '01_Label/'
+            '02_Inspection/'
+        sensor
+            :param
+            '01_IRIS_JX013/'
+            '02_PANDAR_MV/'
+        space
+            :param
+            '01_Highway/'
+            '02_ParkingLot/'
+            '03_Urban/'
+    '''
     step = "02_Inspection/"
     sensor = "01_IRIS_JX013/"
-    space = "03_Urban/"
+    space = "01_Highway/"
 
+    date = '240906'
     print(step, sensor, space)
 
-    drive = "G:/내 드라이브/PM2팀/#김한솔/23-DS-041_현대자동차/3. 데이터/2. 작업 데이터"
+    drive = "C:/Users/pc/SS-233/HDC/"
 
     # STEP 1: Set base file
     try:
         middle_folder_name = os.listdir(f'{drive}/{bucket_name}/{step}/{sensor}/{space}/')[0]
-        data_path = os.path.join(f'{drive}/{bucket_name}/{step}/{sensor}/{space}/', middle_folder_name, '검수 완료 파일 업로드')
+        data_path = os.path.join(f'{drive}/{bucket_name}/{step}/{sensor}/{space}/', middle_folder_name, date)
         inspection_dir = os.path.join(f'{drive}/{bucket_name}/{step}/{sensor}/{space}/', middle_folder_name,
                                       'LDR_GT_Inspection')
 
     except:
         middle_folder_name = os.listdir(f'{drive}/{step}/{sensor}/{space}/')[0]
-        data_path = os.path.join(f'{drive}/{step}/{sensor}/{space}/', middle_folder_name, '검수 완료 파일 업로드')
+        data_path = os.path.join(f'{drive}/{step}/{sensor}/{space}/', middle_folder_name, date)
         inspection_dir = os.path.join(f'{drive}/{step}/{sensor}/{space}/', middle_folder_name, 'LDR_GT_Inspection')
 
     # STEP 2: Process each sequence in the data path
@@ -143,7 +189,7 @@ if __name__ == '__main__':
 
         if sequence.endswith('.ini'):
             continue
-        
+
         print(sequence)
         sequence_path = os.path.join(data_path, sequence)
         save_path = os.path.join(inspection_dir)
